Moved all the harder processing into OrganisationInfo/OrganisationInfo
